relocate package
